Move DataSource class to repository package
