+++ app/src/main/kotlin/com/ayatk/biblio/data/repository/RankingRemoteDataSource.kt
@@ -14,14 +14,13 @@
  * limitations under the License.
  */
 
-package com.ayatk.biblio.data.datasource.ranking
+package com.ayatk.biblio.data.repository
 
 import com.ayatk.biblio.data.narou.NarouClient
 import com.ayatk.biblio.data.narou.entity.enums.NarouRankingType
 import com.ayatk.biblio.data.narou.entity.enums.OutputOrder
 import com.ayatk.biblio.data.narou.entity.mapper.toRanking
 import com.ayatk.biblio.data.narou.util.QueryBuilder
-import com.ayatk.biblio.data.repository.RankingRepository
 import com.ayatk.biblio.model.Ranking
 import com.ayatk.biblio.model.enums.RankingType
 import io.reactivex.Flowable
